[Dockerapi] add pubsub handler for broadcasting in ha setup
